\name{CommandLinePlots-package}
\alias{CommandLinePlots-package}
\alias{CommandLinePlots}
\docType{package}
\title{
Helper library for plot.R
}
\description{
Helper library for plot.R
}
\details{
\tabular{ll}{
Package: \tab CommandLinePlots\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-02-29\cr
License: \tab GPL \cr
}
Helper library for plot.R
}
\author{
Charles Vaske <cvaske@soe.ucsc.edu>

Maintainer: Charles Vaske <cvaske@soe.ucsc.edu>
}
\keyword{ utilities }
